# internal/infrastructure/server/rest/openapi/v1/api.gen.go
// Package v1 provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.7.1 DO NOT EDIT.
package v1

import (
	"encoding/json"
	"errors"
	"fmt"
	"net/http"

	"github.com/go-chi/chi/v5"
	"github.com/oapi-codegen/runtime"
)

// Error defines model for Error.
type Error struct {
	Detail   *string `json:"detail,omitempty"`
	Instance *string `json:"instance,omitempty"`
	Title    string  `json:"title"`
	Type     *string `json:"type,omitempty"`
}

// ErrorResponse defines model for ErrorResponse.
type ErrorResponse struct {
	Error   Error `json:"error"`
	Status  int   `json:"status"`
	Success bool  `json:"success"`
}

// GetProductBySKUResponse defines model for GetProductBySKUResponse.
type GetProductBySKUResponse struct {
	Product Product `json:"product"`
}

// GetSKUsListResponse defines model for GetSKUsListResponse.
type GetSKUsListResponse struct {
	Skus *[]int32 `json:"skus,omitempty"`
}

// Product defines model for Product.
type Product struct {
	Name  *string `json:"Name,omitempty"`
	Price *int32  `json:"Price,omitempty"`
	SKU   *int32  `json:"SKU,omitempty"`
}

// SuccessResponse defines model for SuccessResponse.
type SuccessResponse struct {
	Data    *json.RawMessage `json:"data,omitempty"`
	Status  int              `json:"status"`
	Success bool             `json:"success"`
}

// Sku defines model for sku.
type Sku = int32

// BadRequest defines model for BadRequest.
type BadRequest = ErrorResponse

// InternalServerError defines model for InternalServerError.
type InternalServerError = ErrorResponse

// NotFound defines model for NotFound.
type NotFound = ErrorResponse

// GetProductBySKUParams defines parameters for GetProductBySKU.
type GetProductBySKUParams struct {
	// Sku Артикул продукта (должен быть > 0)
	Sku Sku `form:"sku" json:"sku"`
}

// ListProductSKUsParams defines parameters for ListProductSKUs.
type ListProductSKUsParams struct {
	// StartAfterSku SKU, после которого начинается выборка (должен быть > 0)
	StartAfterSku int32 `form:"start_after_sku" json:"start_after_sku"`

	// Count Количество SKU в ответе (должно быть > 0)
	Count int32 `form:"count" json:"count"`
}

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// Получить продукт по SKU
	// (GET /products/get)
	GetProductBySKU(w http.ResponseWriter, r *http.Request, params GetProductBySKUParams)
	// Получить список SKU
	// (GET /products/list)
	ListProductSKUs(w http.ResponseWriter, r *http.Request, params ListProductSKUsParams)
}

// Unimplemented server implementation that returns http.StatusNotImplemented for each endpoint.

type Unimplemented struct{}

// Получить продукт по SKU
// (GET /products/get)
func (_ Unimplemented) GetProductBySKU(w http.ResponseWriter, r *http.Request, params GetProductBySKUParams) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Получить список SKU
// (GET /products/list)
func (_ Unimplemented) ListProductSKUs(w http.ResponseWriter, r *http.Request, params ListProductSKUsParams) {
	w.WriteHeader(http.StatusNotImplemented)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandlerFunc   func(w http.ResponseWriter, r *http.Request, err error)
}

type MiddlewareFunc func(http.Handler) http.Handler

// GetProductBySKU operation middleware
func (siw *ServerInterfaceWrapper) GetProductBySKU(w http.ResponseWriter, r *http.Request) {

	var err error
	_ = err

	// Parameter object where we will unmarshal all parameters from the context
	var params GetProductBySKUParams

	// ------------- Required query parameter "sku" -------------

	err = runtime.BindQueryParameterWithOptions("form", true, true, "sku", r.URL.Query(), &params.Sku, runtime.BindQueryParameterOptions{Type: "integer", Format: "int32"})
	if err != nil {
		var requiredError *runtime.RequiredParameterError
		if errors.As(err, &requiredError) {
			siw.ErrorHandlerFunc(w, r, &RequiredParamError{ParamName: "sku"})
		} else {
			siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "sku", Err: err})
		}
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetProductBySKU(w, r, params)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// ListProductSKUs operation middleware
func (siw *ServerInterfaceWrapper) ListProductSKUs(w http.ResponseWriter, r *http.Request) {

	var err error
	_ = err

	// Parameter object where we will unmarshal all parameters from the context
	var params ListProductSKUsParams

	// ------------- Required query parameter "start_after_sku" -------------

	err = runtime.BindQueryParameterWithOptions("form", true, true, "start_after_sku", r.URL.Query(), &params.StartAfterSku, runtime.BindQueryParameterOptions{Type: "integer", Format: "int32"})
	if err != nil {
		var requiredError *runtime.RequiredParameterError
		if errors.As(err, &requiredError) {
			siw.ErrorHandlerFunc(w, r, &RequiredParamError{ParamName: "start_after_sku"})
		} else {
			siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "start_after_sku", Err: err})
		}
		return
	}

	// ------------- Required query parameter "count" -------------

	err = runtime.BindQueryParameterWithOptions("form", true, true, "count", r.URL.Query(), &params.Count, runtime.BindQueryParameterOptions{Type: "integer", Format: "int32"})
	if err != nil {
		var requiredError *runtime.RequiredParameterError
		if errors.As(err, &requiredError) {
			siw.ErrorHandlerFunc(w, r, &RequiredParamError{ParamName: "count"})
		} else {
			siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "count", Err: err})
		}
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.ListProductSKUs(w, r, params)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

type UnescapedCookieParamError struct {
	ParamName string
	Err       error
}

func (e *UnescapedCookieParamError) Error() string {
	return fmt.Sprintf("error unescaping cookie parameter '%s'", e.ParamName)
}

func (e *UnescapedCookieParamError) Unwrap() error {
	return e.Err
}

type UnmarshalingParamError struct {
	ParamName string
	Err       error
}

func (e *UnmarshalingParamError) Error() string {
	return fmt.Sprintf("Error unmarshaling parameter %s as JSON: %s", e.ParamName, e.Err.Error())
}

func (e *UnmarshalingParamError) Unwrap() error {
	return e.Err
}

type RequiredParamError struct {
	ParamName string
}

func (e *RequiredParamError) Error() string {
	return fmt.Sprintf("Query argument %s is required, but not found", e.ParamName)
}

type RequiredHeaderError struct {
	ParamName string
	Err       error
}

func (e *RequiredHeaderError) Error() string {
	return fmt.Sprintf("Header parameter %s is required, but not found", e.ParamName)
}

func (e *RequiredHeaderError) Unwrap() error {
	return e.Err
}

type InvalidParamFormatError struct {
	ParamName string
	Err       error
}

func (e *InvalidParamFormatError) Error() string {
	return fmt.Sprintf("Invalid format for parameter %s: %s", e.ParamName, e.Err.Error())
}

func (e *InvalidParamFormatError) Unwrap() error {
	return e.Err
}

type TooManyValuesForParamError struct {
	ParamName string
	Count     int
}

func (e *TooManyValuesForParamError) Error() string {
	return fmt.Sprintf("Expected one value for %s, got %d", e.ParamName, e.Count)
}

// Handler creates http.Handler with routing matching OpenAPI spec.
func Handler(si ServerInterface) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{})
}

type ChiServerOptions struct {
	BaseURL          string
	BaseRouter       chi.Router
	Middlewares      []MiddlewareFunc
	ErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

// HandlerFromMux creates http.Handler with routing matching OpenAPI spec based on the provided mux.
func HandlerFromMux(si ServerInterface, r chi.Router) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseRouter: r,
	})
}

func HandlerFromMuxWithBaseURL(si ServerInterface, r chi.Router, baseURL string) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseURL:    baseURL,
		BaseRouter: r,
	})
}

// HandlerWithOptions creates http.Handler with additional options
func HandlerWithOptions(si ServerInterface, options ChiServerOptions) http.Handler {
	r := options.BaseRouter

	if r == nil {
		r = chi.NewRouter()
	}
	if options.ErrorHandlerFunc == nil {
		options.ErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		}
	}
	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandlerFunc:   options.ErrorHandlerFunc,
	}

	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/products/get", wrapper.GetProductBySKU)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/products/list", wrapper.ListProductSKUs)
	})

	return r
}